+++ R/map_tb_burden.R
@@ -22,8 +22,11 @@
 #' ## Map raw incidence rates
 #' map_tb_burden(download_data = TRUE, save = TRUE)
 #' 
-#' ## Map log scaled incidence rates
-#' map_tb_burden(trans = "log")
+#' ## Map log10 scaled incidence rates
+#' map_tb_burden(trans = "log10")
+#' 
+#' ## Map percentage annual change in incidence rates
+#' map_tb_burden(annual_change = TRUE)
 #' 
 #' ## Find variables relating to mortality in the WHO dataset
 #' search_data_dict(def = "mortality")