+++ tests/testthat/test-get_data.R
@@ -31,3 +31,11 @@ test_that("get_data can download data using utils alternative", {
   
   expect_true(!is.null(tb_data_utils))
 })
+
+
+test_that("get_data fails to download the data when an incorrect URL is given.", {
+  expect_error(get_data(url = NA,
+                        download_data = TRUE,
+                        retry_download = FALSE,
+                        save_name = "test_no_url"))
+})